Need to rework mandatory part to fix some bug

+++ mandatory/include/gfx.h
@@ -16,9 +16,9 @@
 
 int		init_gfx(t_cub3d *c3d);
 void	destroy_gfx(t_graphics *gfx);
-double	get_points_dist(double x1, double y1, double x2, double y2);
-double	deg2rad(double deg);
-double	rad2deg(double rad);
-double	normalize_angle(double angle);
+float	get_points_dist(float x1, float y1, float x2, float y2);
+float	deg2rad(float deg);
+float	rad2deg(float rad);
+float	normalize_angle(float angle);
 
 #endif /* GFX_H */
